requirements.txt
Django==1.11.7
django-admin-sortable2==0.6.17
olefile==0.44
Pillow==4.3.0
pytz==2017.3
django-sendgrid-v5==0.6.85
django-email-obfuscator==0.1.5
